feat(users): use new user entity in handlers

+++ libs/contexts/users/src/lib/users/application/handlers/delete-user-command.handler.ts
@@ -2,12 +2,26 @@ import { BaseError } from '@user-api-monorepo/shared/errors/base.error';
 import { UnexpectedError } from '@user-api-monorepo/shared/errors/unexpected.error';
 import { UserPrimitive } from '../../domain/primitives/user.primitive';
 import { UserRepository } from '../../domain/repositories/user.repository';
+import { DatabaseRecordNotFoundError } from '@user-api-monorepo/shared/errors/database-record-not-found.error';
+import { UserEntity } from '../../domain/entities/user.entity';
 
 export class DeleteUserCommandHandler {
   constructor(private readonly userRepository: UserRepository) {}
 
   async execute(id: UserPrimitive['id']): Promise<void> {
     try {
+      const userPrimitive = await this.userRepository.getUser(id);
+
+      if (!userPrimitive) {
+        throw new DatabaseRecordNotFoundError(
+          this.constructor.name,
+          'execute',
+          id,
+        );
+      }
+
+      UserEntity.fromPrimitives(userPrimitive);
+
       return this.userRepository.deleteUser(id);
     } catch (error) {
       if (error instanceof BaseError) {
